refactor: start splitting to files
